Changed the save/load system
There used to be 2 different ways the data in AstroRun got saved. The first being on the device itself using PlayerPrefs. This makes the user non-dependent on a network connection. But has the disadvantage of being tied to the device. So if the user would switch devices, all progress would be lost. To counteract this, AstroRun made use of the Google Play Games API to save data to the user's google account. This would allow the user to play from any device without losing progress. However, the solution using the  Play games API didn't seem to work.
So instead we opted to use a local save file along side playerprefs. This is way more reliable then the Play Games API.  Whilest still allowing the user to carry over it's progress to another device. This is achieved by simple copying the save file from one of the devices to the other.