Reduced the overall difficulty of level 7
Increased the distance seperating the 3 'vallende spikes' in the beginning of the level

Increased the jump strength of the 'spring' from 1000 to 1100. The 'spring' that's located just before the 'vallende spike' that was hidden behind the waterfall.

Removed the waterfall that was in front of the 'vallende spike'

Reduced the size of the first 'Saw' by 30%. From 1 1 1 -> 0.7 0.7 0.7