Fixed an issue that a new personal best wouldn't be submitted to the leaderboards
If the user beats his personal best on a certain level, the game will try to upload it to the leaderboards upon finishing the level. If it fails, the time gets saved to upload it at a later time.

The game would try a second time when opening the level selection screen on the main menu. If it would fail again, this new personal best would be lost.

Now the game will keep trying to upload the new personal best, until it succeeds. This will happen every time the level selection menu is opened